Also walk `target.'cfg()'` for `dev-dependencies` next to `dependencies`
In [RustAudio/cpal#641] we're seeing a [failure in `ndk-macro`] to turn
a crate name with dashes into underscores as fallback, with the
underlying issue that `proc-macro-crate` failed to find our crate in
`Cargo.toml` when limited to a specific target cfg.  According to [the
Rust docs] it is perfectly possible to have `dev-dependencies` behind
such a target predicate too.

[RustAudio/cpal#641]: https://github.com/RustAudio/cpal/pull/641#issuecomment-1040470392
[failure in `ndk-macro`]: https://github.com/rust-windowing/android-ndk-rs/blob/9060d5dbea8bdd9aee9e42ef8e8375c35adbc271/ndk-macro/src/helper.rs#L64
[the Rust docs]: https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#development-dependencies